+++ api/build.gradle
@@ -2,7 +2,11 @@ bootJar { enabled = true }
 jar { enabled = false }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     implementation project(':domain')
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    // swagger
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
