+++ api/src/main/java/com/tago/api/auth/api/AuthApi.java
@@ -4,6 +4,7 @@ package com.tago.api.auth.api;
 import com.tago.api.auth.application.OAuthLoginService;
 import com.tago.api.auth.dto.LoginResponse;
 import com.tago.api.auth.infra.kakao.KakaoLoginParams;
+import com.tago.api.common.model.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,13 +14,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/auth")
+@RequestMapping("/api/v1/")
 public class AuthApi {
     private final OAuthLoginService oAuthLoginService;
 
-    @PostMapping("/kakao")
+    @PostMapping("/auth/kakao")
     public ResponseEntity<LoginResponse> loginKakao(@RequestBody KakaoLoginParams params) {
-        return ResponseEntity.ok(oAuthLoginService.login(params));
+        return ResponseDto.ok(oAuthLoginService.login(params));
     }
 
 }
