+++ api/src/main/java/com/tago/api/auth/infra/kakao/KakaoApiClient.java
@@ -1,9 +1,9 @@
 package com.tago.api.auth.infra.kakao;
 
-import com.tago.domain.auth.model.OAuthApiClient;
-import com.tago.domain.auth.model.OAuthInfoResponse;
-import com.tago.domain.auth.model.OAuthLoginParams;
-import com.tago.domain.auth.model.OAuthProvider;
+import com.tago.domain.auth.domain.OAuthApiClient;
+import com.tago.domain.auth.domain.OAuthInfoResponse;
+import com.tago.domain.auth.domain.OAuthLoginParams;
+import com.tago.domain.auth.domain.OAuthProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
