+++ api/src/main/java/com/tago/api/auth/infra/kakao/KakaoInfoResponse.java
@@ -1,7 +1,7 @@
 package com.tago.api.auth.infra.kakao;
 
-import com.tago.domain.auth.model.OAuthInfoResponse;
-import com.tago.domain.auth.model.OAuthProvider;
+import com.tago.domain.auth.domain.OAuthInfoResponse;
+import com.tago.domain.auth.domain.OAuthProvider;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
