--- api/src/main/java/com/tago/api/config/ClientConfig.java
@@ -1,14 +1,0 @@
-package com.tago.api.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class ClientConfig {
-
-    @Bean
-    public RestTemplate restTemplate(){
-        return new RestTemplate();
-    }
-}
