+++ domain/src/main/java/com/tago/domain/auth/domain/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.tago.domain.auth.model;
+package com.tago.domain.auth.domain;
 
 public enum OAuthProvider {
